Separación del proyecto en tres módulos: imc, auth y users. Los tres funcionan correctamente, pero aún no se encuentran integrados
