+++ src/app.module.ts
@@ -3,11 +3,12 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 
-import { CalculoImc } from './module/imc/entities/CalculoImc.entity';
-import { ImcModule } from './module/imc/imc.module';
+import { CalculoImc } from './imc/entities/CalculoImc.entity';
+import { ImcModule } from './imc/imc.module';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { UsersModule } from './auth/users/users.module';
+import { UsersModule } from './users/users.module';
+import { AuthModule } from './auth/auth.module';
 
 @Module({
   imports: [
@@ -33,6 +34,7 @@ import { UsersModule } from './auth/users/users.module';
 
     ImcModule,
     UsersModule,
+    AuthModule,
   ],
   controllers: [AppController],
   providers: [AppService],
