+++ src/auth/guards/auth.guard.ts
@@ -1,7 +1,7 @@
 //ARCHIVO: auth.middleware.ts
 import { CanActivate, ExecutionContext, Injectable, UnauthorizedException } from "@nestjs/common";
 import { JwtService } from "../jwt/jwt.service";
-import { UsersService } from "../users/users.service";
+import { UsersService } from "../../users/users.service";
 import { RequestWithUser } from "../interfaces/request-with-user.interface";
 
 /*
