+++ src/imc/imc.service.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from "@nestjs/testing";
 import { ImcService } from "./imc.service";
-import { CalcularImcDto } from "./dto/calculo-imc.dto";
+import { CalculoImcDto } from "./dto/calculo-imc.dto";
 
 
 describe('ImcService', () => {
@@ -19,28 +19,28 @@ describe('ImcService', () => {
   });
 
   it('should calculate IMC correctly', () => {
-    const dto: CalcularImcDto = { altura: 1.75, peso: 70 };
+    const dto: CalculoImcDto = { altura: 1.75, peso: 70 };
     const result = service.calcularImc(dto);
     expect(result.imc).toBeCloseTo(22.86, 2); // Redondeado a 2 decimales
     expect(result.categoria).toBe('Normal');
   });
 
   it('should return Bajo peso for IMC < 18.5', () => {
-    const dto: CalcularImcDto = { altura: 1.75, peso: 50 };
+    const dto: CalculoImcDto = { altura: 1.75, peso: 50 };
     const result = service.calcularImc(dto);
     expect(result.imc).toBeCloseTo(16.33, 2);
     expect(result.categoria).toBe('Bajo peso');
   });
 
   it('should return Sobrepeso for 25 <= IMC < 30', () => {
-    const dto: CalcularImcDto = { altura: 1.75, peso: 80 };
+    const dto: CalculoImcDto = { altura: 1.75, peso: 80 };
     const result = service.calcularImc(dto);
     expect(result.imc).toBeCloseTo(26.12, 2);
     expect(result.categoria).toBe('Sobrepeso');
   });
 
   it('should return Obeso for IMC >= 30', () => {
-    const dto: CalcularImcDto = { altura: 1.75, peso: 100 };
+    const dto: CalculoImcDto = { altura: 1.75, peso: 100 };
     const result = service.calcularImc(dto);
     expect(result.imc).toBeCloseTo(32.65, 2);
     expect(result.categoria).toBe('Obeso');
